Use browser history to avoid hash in URL
Uses the browser history package instead of the hash one that puts ugly
looking hashes in the URL to maintain state. There is only one problem
with doing this: you can no longer access the website directly through a
link to a subpage. That is, https://yourapp.example.xyz/ will serve the
front page of the app (and let you navigate to subpages such as
/molecule/Molecule3), but you cannot paste that subpath into the browser
URL and expect a response.

Solving this would require a simple server that _always_ serves
index.html regardless of which subpath is requested.